Switch server from Hypercorn to Uvicorn in Dockerfile
This commit changes the server used in Dockerfile from Hypercorn to Uvicorn, aligning the server used in production with the one used during development for consistency.

It also includes the addition of a .gitignore file to prevent unnecessary files from being tracked and updates FastAPI, Trio, Hypercorn, OpenAI and HTTPX to use the latest versions. Plus, the Uvicorn and Gunicorn libraries were added to the project's dependencies.