Update default chatbot message and remove ngrok-skip-browser-warning response header
With this commit, we change the default bot message to have a more friendly introductory message and provide some suggestion buttons to guide the user's interaction. This aims to improve user experience. The "ngrok-skip-browser-warning" header in the 'read_root' function has removed since it's not necessary anymore for the functionality of the app.